# docs/swagger.yaml
definitions:
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  handlers.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  handlers.LoginResponse:
    properties:
      needOnboarding:
        type: boolean
      token:
        type: string
      user:
        $ref: '#/definitions/models.User'
    type: object
  handlers.OnboardingResponse:
    properties:
      message:
        type: string
    type: object
  handlers.UserResponse:
    properties:
      email:
        type: string
      user_id:
        type: integer
    type: object
  models.Company:
    properties:
      address:
        type: string
      alley:
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      district:
        type: string
      id:
        type: integer
      name:
        type: string
      postal_code:
        type: string
      province:
        type: string
      sub_district:
        type: string
      tel:
        type: string
      type:
        type: string
      updatedAt:
        type: string
    type: object
  models.OnboardingPayload:
    properties:
      address:
        example: 123 ถนนสุขุมวิท
        type: string
      alley:
        example: Soi 5
        type: string
      companyName:
        example: ABC Co.
        type: string
      companyType:
        example: SME
        type: string
      district:
        example: Pathum Wan
        type: string
      postalCode:
        example: "10330"
        type: string
      province:
        example: Bangkok
        type: string
      subDistrict:
        example: Lumphini
        type: string
      tel:
        example: "0999999999"
        type: string
    type: object
  models.User:
    properties:
      company:
        $ref: '#/definitions/models.Company'
      companyID:
        type: integer
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      email:
        type: string
      google_id:
        type: string
      id:
        type: integer
      is_validated:
        type: boolean
      name:
        type: string
      tel:
        type: string
      updatedAt:
        type: string
    type: object
info:
  contact: {}
paths:
  /api/auth/google:
    post:
      consumes:
      - application/json
      description: Verify Google ID Token and return JWT
      parameters:
      - description: Google ID Token
        in: body
        name: body
        required: true
        schema:
          properties:
            id_token:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Login with Google
      tags:
      - auth
  /api/onboarding:
    post:
      consumes:
      - application/json
      description: Save company and user info (step before OTP)
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Onboarding data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.OnboardingPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.OnboardingResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: User onboarding
      tags:
      - user
  /api/test:
    get:
      description: Get user_id from JWT
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.UserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - user
swagger: "2.0"